requirements.txt
discord.py==2.3.2
requests==2.32.0
aiohttp==3.12.15
Flask==3.0.0
aiohttp==3.12.15
discord.py==2.3.2
Flask==3.0.0
requests==2.32.0
